requirements.txt
langchain
chromadb
dspy-ai
openai
anthropic
faiss-cpu